changed class name to a more generic one in order to have all security headers included
