--- src/main/java/com/systelab/seed/infrastructure/security/CORSFilter.java
@@ -1,23 +1,0 @@
-package com.systelab.seed.infrastructure.security;
-
-import java.io.IOException;
-
-import javax.ws.rs.container.ContainerRequestContext;
-import javax.ws.rs.container.ContainerResponseContext;
-import javax.ws.rs.container.ContainerResponseFilter;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-public class CORSFilter implements ContainerResponseFilter {
-
-    @Override
-    public void filter(ContainerRequestContext request, ContainerResponseContext response) throws IOException {
-        response.getHeaders().add("Access-Control-Allow-Origin", "*");
-        response.getHeaders().add("Access-Control-Allow-Headers", "origin, content-type, accept, authorization, Etag, if-none-match");
-        response.getHeaders().add("Access-Control-Allow-Credentials", "true");
-        response.getHeaders().add("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS, HEAD");
-        response.getHeaders().add("Access-Control-Max-Age", "1209600");
-        response.getHeaders().add("Access-Control-Expose-Headers", "origin, content-type, accept, authorization, ETag, if-none-match");
-        response.getHeaders().add("X-XSS-Protection", "1; mode=block");
-    }
-}
